pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["TamirEdelstein <tamir4989@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.53.0"
pandas-stubs = "~=2.3.3"
plotly = "^6.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
